requirements.txt
Django==2.2.5
django-environ==0.4.5
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0
lxml==4.4.1
numpy==1.17.2
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
regex==2019.8.19
six==1.12.0
sqlparse==0.3.0
